pyproject.toml
[tool.poetry]
name = "dataservices"
version = "0.1.0"
description = ""
authors = ["Sean Baier <sean.t.baier@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.19.7"
loguru = "^0.5.3"
python-dotenv = "^0.19.1"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
isort = "^5.9.3"
mypy = "^0.910"
flake8 = "^4.0.1"
pytest = "^6.2.5"
pytest-watch = "^4.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
"ls:build" = "scripts.infrastructure.build:build"
"ls:destroy" = "scripts.infrastructure.destroy:destroy"
clean = "scripts.utils:clean"
build = "scripts.utils:build"